+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2037** : Add config prop to ensure every REST call is logged
+
 * Issue **#2038** : Allow autologger action to be modified (search and process)
 
 * Issue **#2027** : Fix autologger update operation